ChangesV1

+++ Dtos/UserDto.cs
@@ -5,6 +5,6 @@ namespace AcademyTest.Dtos
         public Guid Id { get; init; }
         public string Name { get; init; }
         public string Email { get; init; }
-        public int Phone { get; init; }
+        public long Phone { get; init; }
     }
 }
